% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_10Reviews.R
\name{get_10Reviews}
\alias{get_10Reviews}
\title{Extracts all components of a review such as review title, date, location, device, etc.}
\usage{
get_10Reviews(node)
}
\arguments{
\item{node}{the single review given by a user}
}
\description{
Extracts all components of a review such as review title, date, location, device, etc.
}
\references{
Inspired by Hadley Wickham (author of pkg::rvest) and BjaRule (StackOverflow).
https://github.com/hadley/rvest/issues/12
https://stackoverflow.com/questions/33250826/scraping-with-rvest-complete-with-nas-when-tag-is-not-present
}
\seealso{
The main program get_allReviews.R make use of this program to extract all the components of a given review.
}
\author{
Awash Analytics (December 2017)
}
