% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_allReviews.R
\name{get_allReviews}
\alias{get_allReviews}
\title{Extracts reviews of an attraction site posted by tourists on TripAdvisor website.
It takes care of both a single page and multiple pages.}
\usage{
get_allReviews(nbr_page = NULL, site_gCode = NULL, site_dCode = NULL,
  site_name = NULL, site_geoLocation = NULL)
}
\arguments{
\item{nbr_page}{the number of pages to grab.}

\item{site_gCode}{the "g" code as defined by TripAdvisor pagination. Please see the example below.}

\item{site_dCode}{the "d" code as defined by TripAdvisor pagination. Please see the example below.

Extracts reviews of an attraction site posted by tourists on TripAdvisor website.
It takes care of both a single page and multiple pages.}

\item{site_name}{the name of an attraction site.}

\item{site_geoLocation}{the geographical location of an attraction site. Not that this is not related to a latitude/longitude values. Please see the example below.}

\item{nbr_page}{the number of pages to grab.}

\item{site_gCode}{the "g" code as defined by TripAdvisor pagination. Please see the example below.}

\item{site_dCode}{the "d" code as defined by TripAdvisor pagination. Please see the example below.}
}
\description{
This source code calls get_10Reviews.R program to extract the required components (e.g., review title, date, device). Depending on
a user request, it can also loop through multiple pages. Finally, it adds extra columns (e.g., attraction location, number of reviews, etc)
to existing reviews stored in a dataframe.
}
\examples{
## reviews_lalibela <- get_allReviews(nbr_page = 2,
##                                    site_gCode = "480193",
##                                    site_dCode = "324957",
##                                    site_name = "Rock_Hewn_Churches_of_Lalibela",
##                                    site_geoLocation = "Lalibela_Amhara_Region"
##                                    )

}
\author{
Awash Analytics (December 2017)
}
